charlib/characterizer: Move Port into its own module, filter StateFunction.test_vectors by trigger activity, and add a regression test for StateFunctions
