+++ pyproject.toml
@@ -41,4 +41,7 @@ requires = ["poetry-core>=2.0.0,<3.0.0"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.pytest.ini_options]
-testpaths = ["test/syntax"]
+testpaths = [
+    "test/syntax",
+    "test/logic"
+]
